Script cleanup/refactor
- remove unnecessary script
- rename script (semantic accuracy)
