+++ package.json
@@ -8,8 +8,7 @@
     "db:reset": "npm run db drop && npm run db create && npm run db migrate",
     "db:seed": "npm run db seed:all",
     "add:integration": "node  -r dotenv/config ./server/script/add-integration.js",
-    "duration:aggregate": "node  -r dotenv/config ./server/script/interaction-duration.js aggregate",
-    "duration:remove": "node  -r dotenv/config ./server/script/interaction-duration.js remove",
+    "profiles:aggregate-stats": "node  -r dotenv/config ./server/script/aggregate-stats.js",
     "lint": "eslint --ext .js server",
     "debug": "nodemon --inspect-brk -r dotenv/config ./server | bunyan",
     "start": "NODE_ENV=production node -r dotenv/config ./server",
